[codex] Fix CodeNarc violations in matrix-ggplot

+++ matrix-ggplot/src/test/groovy/gg/scale/ScaleAlphaIdentityTest.groovy
@@ -17,37 +17,37 @@ class ScaleAlphaIdentityTest {
   @Test
   void testBasicTransform() {
     def scale = new ScaleAlphaIdentity()
-    assertEquals(0.5 as BigDecimal, scale.transform(0.5))
-    assertEquals(0.8 as BigDecimal, scale.transform(0.8))
-    assertEquals(1.0 as BigDecimal, scale.transform(1.0))
+    assertEquals(0.5, scale.transform(0.5))
+    assertEquals(0.8, scale.transform(0.8))
+    assertEquals(1.0, scale.transform(1.0))
   }
 
   @Test
   void testClampingUpperBound() {
     def scale = new ScaleAlphaIdentity()
-    assertEquals(1.0 as BigDecimal, scale.transform(1.5))
-    assertEquals(1.0 as BigDecimal, scale.transform(2.0))
-    assertEquals(1.0 as BigDecimal, scale.transform(100))
+    assertEquals(1.0, scale.transform(1.5))
+    assertEquals(1.0, scale.transform(2.0))
+    assertEquals(1.0, scale.transform(100))
   }
 
   @Test
   void testClampingLowerBound() {
     def scale = new ScaleAlphaIdentity()
-    assertEquals(0.0 as BigDecimal, scale.transform(-0.5))
-    assertEquals(0.0 as BigDecimal, scale.transform(-1.0))
-    assertEquals(0.0 as BigDecimal, scale.transform(-100))
+    assertEquals(0.0, scale.transform(-0.5))
+    assertEquals(0.0, scale.transform(-1.0))
+    assertEquals(0.0, scale.transform(-100))
   }
 
   @Test
   void testZeroAlpha() {
     def scale = new ScaleAlphaIdentity()
-    assertEquals(0.0 as BigDecimal, scale.transform(0.0))
+    assertEquals(0.0, scale.transform(0.0))
   }
 
   @Test
   void testOneAlpha() {
     def scale = new ScaleAlphaIdentity()
-    assertEquals(1.0 as BigDecimal, scale.transform(1.0))
+    assertEquals(1.0, scale.transform(1.0))
   }
 
   @Test
@@ -55,6 +55,6 @@ class ScaleAlphaIdentityTest {
     def scale = new ScaleAlphaIdentity()
     def result = scale.transform('0.6')
     assertNotNull(result)
-    assertEquals(0.6 as BigDecimal, result)
+    assertEquals(0.6, result)
   }
 }
